+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.18.11] - 2026-04-09
+
 ### Changed
 
 - Stream chunks are written to stdout incrementally instead of buffering until drain completes